requirements.txt
# Core PyTorch and Vision Libraries
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Image Processing
Pillow>=9.0.0

# Scientific Computing and Statistics
numpy>=1.21.0
scipy>=1.7.0

# Progress Bars and CLI
tqdm>=4.64.0

# Configuration File Support
PyYAML>=6.0

# Testing (Development)
pytest>=7.0.0
pytest-mock>=3.10.0

# Additional Metrics
lpips>=0.1.4

# Optional: For visualization (uncomment if needed)
# matplotlib>=3.5.0
# seaborn>=0.11.0

# Optional: For advanced FID computation
# pytorch-fid>=0.3.0

# Development Tools (uncomment for development)
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950
